pyproject.toml
[project]
name = "farcaster-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.10.0",
    "farcaster>=0.7.12",
    "requests>=2.32.4",
    "starlette>=0.47.2",
    "urllib3>=2.5.0",
    "python-dotenv>=1.0.1",
]
